Step 8

+++ src/services/contacts.js
@@ -2,11 +2,11 @@ import { ContactsCollection } from '../db/models/contacts.js';
 import { calculatePaginationData } from '../utils/calculatePaginationData.js';
 import { SORT_ORDER } from '../constants/index.js';
 
-export const getAllContacts = async ({ page, perPage, sortBy = 'name', sortOrder = SORT_ORDER.ASC, filter = {} }) => {
+export const getAllContacts = async ({ userId, page, perPage, sortBy = 'name', sortOrder = SORT_ORDER.ASC, filter = {} }) => {
     const limit = perPage;
     const skip = (page - 1) * perPage;
 
-    const contactsQuery = ContactsCollection.find();
+    const contactsQuery = ContactsCollection.find({ userId });
 
     if (filter.type) {
         contactsQuery.where('contactType').equals(filter.type);
@@ -32,8 +32,8 @@ export const getAllContacts = async ({ page, perPage, sortBy = 'name', sortOrder
     };
 };
 
-export const getContactById = async (contactId) => {
-    const contact = await ContactsCollection.findById(contactId);
+export const getContactById = async (userId, contactId) => {
+    const contact = await ContactsCollection.findOne({ _id: contactId, userId });
     return contact;
 };
 
@@ -42,14 +42,14 @@ export const createContact = async (payload) => {
     return newContact;
 };
 
-export const deleteContact = async (contactId) => {
-    const deletedContact = await ContactsCollection.findOneAndDelete({ _id: contactId });
+export const deleteContact = async (userId, contactId) => {
+    const deletedContact = await ContactsCollection.findOneAndDelete({ _id: contactId, userId });
     return deletedContact;
 };
 
-export const updateContact = async (contactId, payload, options = {}) => {
+export const updateContact = async (userId, contactId, payload, options = {}) => {
     const updatedContact = await ContactsCollection.findOneAndUpdate(
-        { _id: contactId },
+        { _id: contactId, userId },
         payload,
         {
             new: true,
